:sparkles: Feat: about 애니메이션 속도 조절
그렇다...
